add data migrations for districts and related fields
